requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
Django==2.1.2
djangorestframework==3.9.0
kombu==4.2.1
more-itertools==4.3.0
nltk==3.3
numpy==1.15.3
pkg-resources==0.0.0
psycopg2==2.7.5
python-dateutil==2.7.5
pytz==2018.6
redis==2.10.6
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
vine==1.1.4
